+++ fractal/ComponentRenderer.jsx
@@ -1,9 +1,10 @@
 import React from 'react';
 import ReactDOMServer from 'react-dom/server';
 
-import SVGIcon from './components/icons/SVGIcon/SVGIcon.jsx';
+import DocumentHead from './components/page-sections/DocumentHead/DocumentHead.jsx';
 import SiteHeader from './components/page-sections/SiteHeader/SiteHeader.jsx';
 import SiteFooter from './components/page-sections/SiteFooter/SiteFooter.jsx';
+import SVGIcon from './components/icons/SVGIcon/SVGIcon.jsx';
 
 import SVGIconCSS from './components/icons/SVGIcon/SVGIcon.css';
 import ThemeESILightSeaGreenCSS from './scss/motifs/light-sea-green/light-sea-green-esi.css';
@@ -14,6 +15,7 @@ class ComponentRenderer {
 
     static getComponents() {
         return {
+            DocumentHead,
             SiteFooter,
             SiteHeader,
             SVGIcon