+++ fractal/components/page-sections/PageTemplate/PageTemplate.jsx
@@ -1,6 +1,7 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 
+import DocumentHead from '../DocumentHead/DocumentHead.jsx';
 import PageLayer from '../../containers/PageLayer/PageLayer.jsx';
 import SiteHeader from '../SiteHeader/SiteHeader.jsx';
 import SiteFooter from '../SiteFooter/SiteFooter.jsx';
@@ -15,16 +16,22 @@ const PageTemplate = (props) => {
     // Extend class names
 
     // Compose props
+    const documentHeadProps = {};
+
+
 
     // Render
     return (
-        <div>
-            <SiteHeader {...props} className={esiClass}/>
-            <main>
-                <PageLayer>XXX page content XXX</PageLayer>
-            </main>
-            <SiteFooter className={esiClass}/>
-        </div>
+        <html lang="en" xmlns="" className="">
+            <DocumentHead {...documentHeadProps}/>
+            <body>
+                <SiteHeader {...props} className={esiClass}/>
+                <main>
+                    <PageLayer>XXX page content XXX</PageLayer>
+                </main>
+                <SiteFooter className={esiClass}/>
+            </body>
+        </html>
     );
 };