+++ fractal/components/page-sections/_document-head-preview.jsx
@@ -1,0 +1,13 @@
+import React from 'react';
+
+const page = ({ rendered_content, path }) => {
+
+    return (
+        <div>
+            <script src="http://code.jquery.com/jquery-3.2.1.min.js" integrity="sha256-hwg4gsxgFZhOsEEamdOYGBf13FyQuiTwlAQgxVSNgt4=" crossOrigin="anonymous"/>
+            <div className="base-preview" dangerouslySetInnerHTML={{__html: rendered_content}}/>
+        </div>
+    );
+};
+
+module.exports = page;